requirements.txt
streamlit
pyrebase4